requirements.txt
APscheduler
colorama
pyfiglet
python-dotenv
requests
